refaktorert skjema
